refactor a lot of stuff
