+++ src/lib/components/course/index.d.ts
@@ -1,0 +1,5 @@
+export type FileStructure = {
+	type: 'folder' | 'file';
+	name: string;
+	files?: FileStructure[];
+};
