+++ src/lib/components/icons/Hamburger.svelte
@@ -1,0 +1,13 @@
+<svg
+	xmlns="http://www.w3.org/2000/svg"
+	class="h-5 w-5"
+	fill="none"
+	viewBox="0 0 24 24"
+	stroke="currentColor"
+	><path
+		stroke-linecap="round"
+		stroke-linejoin="round"
+		stroke-width="2"
+		d="M4 6h16M4 12h8m-8 6h16"
+	/></svg
+>
