+++ src/lib/components/icons/index.ts
@@ -1,0 +1,6 @@
+import Github from "./Github.svelte";
+import Gmail from "./Gmail.svelte";
+import LinkedIn from "./LinkedIn.svelte";
+import Hamburger from "./Hamburger.svelte";
+
+export { Github, Gmail, LinkedIn, Hamburger };
