+++ src/lib/components/index.ts
@@ -1,8 +1,8 @@
 // Reexport your entry components here
 import Intro from './intro/Intro.svelte';
 import Work from './work/Work.svelte';
-import PersonalProject from './project/personal/PersonalProject.svelte';
-import SchoolProject from './project/school/SchoolProject.svelte';
+import Project from './project/PersonalProject.svelte';
+import Course from './course/Course.svelte';
 import Skill from './skill/skill.svelte';
 
-export { Intro, Work, PersonalProject, SchoolProject, Skill };
+export { Intro, Work, Project, Course, Skill };
