+++ src/routes/Navbar.svelte
@@ -1,148 +1,60 @@
-<div class="flex justify-between items-center pt-2 pl-2">
-	<div class="pt-1 pb-2">
-		<h1 class="text-2xl font-light font-serif">
-			<a href="/">Yongye Tan </a>
-		</h1>
+<script lang="ts">
+	import { Github, LinkedIn, Gmail, Hamburger } from '$lib/components/icons';
+</script>
+
+<div class="flex justify-between items-center navbar pt-2 pl-2">
+	<div class="navbar-start">
+		<div class="dropdown">
+			<div tabindex="0" role="button" class="btn btn-ghost lg:hidden">
+				<Hamburger />
+			</div>
+			<ul class="menu menu-sm dropdown-content mt-3 z-[1] p-2 bg-base-100 rounded-box w-52">
+				<li><a href="/projects">Projects</a></li>
+				<li><a href="/courses">Courses</a></li>
+			</ul>
+		</div>
+		<div class="pt-1 pb-2">
+			<a class="text-2xl font-light font-serif" href="/">Yongye Tan</a>
+		</div>
 	</div>
-
-	<div class="flex flex-wrap gap-1">
-		<div class="relative group">
-			<a href="mailto:yongyetan1209@gmail.com" class="relative z-10" title="email">
-				<svg
-					xmlns="http://www.w3.org/2000/svg"
-					x="0px"
-					y="0px"
-					width="24"
-					height="24"
-					viewBox="0 0 100 100"
-				>
-					<polygon
-						fill="#f16e7c"
-						points="29.434,25.293 50,40.657 70.566,25.293 70.566,48 50,63 29.434,48"
-					/><path
-						fill="#ead032"
-						d="M70.566,47.554L87.91,34.849v-4.808c0-5.002-4.055-9.056-9.056-9.056h-2.046l-6.242,4.308	C70.566,25.293,70.566,47.554,70.566,47.554z"
-					/><path
-						fill="#e85654"
-						d="M29.434,47.554L12.089,34.849v-4.808c0-5.002,4.055-9.056,9.056-9.056h2.046l6.242,4.308	L29.434,47.554L29.434,47.554z"
-					/><path
-						fill="#8cc78c"
-						d="M70,48l18.362-13.638v38.926c0,3.155-2.557,5.712-5.712,5.712H70V48z"
-					/><path
-						fill="#40a6dd"
-						d="M30,48L11.638,34.362v38.926c0,3.155,2.557,5.712,5.712,5.712H30V48z"
-					/><path
-						fill="#1f212b"
-						d="M78.99,20c-2.186,0-4.265,0.695-6.012,2.011L50,39.328L27.021,22.011C25.274,20.695,23.195,20,21,20	c-5.514,0-10,4.486-10,10v44c0,3.309,2.691,6,6,6h13c0.553,0,1-0.447,1-1V50.092L49.4,63.8c0.18,0.13,0.39,0.2,0.6,0.2	s0.41-0.06,0.59-0.19L69,50.514V79c0,0.553,0.447,1,1,1h13c3.309,0,6-2.691,6-6V30C89,24.486,84.514,20,78.99,20z M74.182,23.608	C75.579,22.557,77.242,22,79,22c4.411,0,8,3.589,8,8v4.799L71,46.724V46.6V26.006L74.182,23.608z M21.01,22	c1.748,0,3.411,0.557,4.809,1.608L29,26.006V46.6v0.124L13,34.799V30C13,25.589,16.589,22,21.01,22z M29,78H17c-2.206,0-4-1.794-4-4	V36.047l16,11.924v0.823V49V78z M50.01,61.76L30,47.322V26.76l19.398,14.619c0.017,0.013,0.037,0.015,0.054,0.026	c0.059,0.039,0.122,0.064,0.187,0.089c0.057,0.022,0.111,0.047,0.17,0.059c0.066,0.013,0.13,0.011,0.197,0.01	c0.061,0,0.12,0.002,0.18-0.009c0.062-0.012,0.12-0.038,0.18-0.062c0.063-0.025,0.124-0.049,0.181-0.087	c0.017-0.011,0.037-0.013,0.054-0.026L70,26.76v20.562L50.01,61.76z M83,78H71V49.07v-0.174V48.5v-0.528l16-11.924V74	C87,76.206,85.206,78,83,78z"
-					/><path
-						fill="#1f212b"
-						d="M16.5,61c0.276,0,0.5-0.224,0.5-0.5v-2c0-0.276-0.224-0.5-0.5-0.5S16,58.224,16,58.5v2	C16,60.776,16.224,61,16.5,61z"
-					/><path
-						fill="#1f212b"
-						d="M16.5,52c0.276,0,0.5-0.224,0.5-0.5v-7c0-0.276-0.224-0.5-0.5-0.5S16,44.224,16,44.5v7	C16,51.776,16.224,52,16.5,52z"
-					/><path
-						fill="#1f212b"
-						d="M24.5,74h-7c-0.275,0-0.5-0.225-0.5-0.5v-10c0-0.276-0.224-0.5-0.5-0.5S16,63.224,16,63.5v10	c0,0.827,0.673,1.5,1.5,1.5h7c0.276,0,0.5-0.224,0.5-0.5S24.776,74,24.5,74z"
-					/><path
-						fill="#1f212b"
-						d="M42.801,40.16c-0.22-0.164-0.532-0.123-0.7,0.099c-0.166,0.221-0.122,0.534,0.099,0.7l2.047,1.542	c0.09,0.067,0.195,0.101,0.301,0.101c0.151,0,0.301-0.068,0.399-0.199c0.166-0.221,0.122-0.534-0.099-0.7L42.801,40.16z"
-					/><path
-						fill="#1f212b"
-						d="M46.444,42.905c-0.219-0.164-0.533-0.123-0.7,0.099c-0.166,0.221-0.122,0.534,0.099,0.7l2.914,2.195	C48.847,45.967,48.952,46,49.058,46c0.151,0,0.301-0.068,0.399-0.199c0.166-0.221,0.122-0.534-0.099-0.7L46.444,42.905z"
-					/><path
-						fill="#1f212b"
-						d="M55.306,41.587l-3.337,2.514c-0.221,0.166-0.265,0.479-0.099,0.7C51.969,44.932,52.118,45,52.27,45	c0.105,0,0.211-0.033,0.301-0.101l3.337-2.514c0.221-0.166,0.265-0.479,0.099-0.7C55.839,41.464,55.524,41.423,55.306,41.587z"
-					/><path
-						fill="#1f212b"
-						d="M58.682,39.044l-1.778,1.339c-0.221,0.166-0.265,0.479-0.099,0.7	c0.099,0.131,0.248,0.199,0.399,0.199c0.105,0,0.211-0.033,0.301-0.101l1.778-1.339c0.221-0.166,0.265-0.479,0.099-0.7	C59.215,38.92,58.9,38.88,58.682,39.044z"
-					/><path
-						fill="#1f212b"
-						d="M66.899,33.479c-0.167-0.222-0.481-0.263-0.7-0.099l-5.919,4.46c-0.221,0.166-0.265,0.479-0.099,0.7	c0.099,0.131,0.248,0.199,0.399,0.199c0.105,0,0.211-0.033,0.301-0.101l5.919-4.459C67.021,34.014,67.065,33.7,66.899,33.479z"
-					/><path
-						fill="#1f212b"
-						d="M16.5,56c0.276,0,0.5-0.224,0.5-0.5v-1c0-0.276-0.224-0.5-0.5-0.5S16,54.224,16,54.5v1	C16,55.776,16.224,56,16.5,56z"
-					/>
-				</svg>
-			</a>
-			<div class="absolute top-1 -right-1 group-hover:block pt-6 pr-1 hidden">
-				<div class="bg-gray-200 px-2 py-0.5">
-					<p class="text-sm font-medium">yongyetan1209@gmail.com</p>
+	<div class="navbar-center hidden lg:flex">
+		<ul class="menu menu-horizontal px-1">
+			<li><a href="/projects">Projects</a></li>
+			<li><a href="/courses">Courses</a></li>
+		</ul>
+	</div>
+	<div class="navbar-end">
+		<div class="flex flex-wrap gap-1">
+			<div class="relative group">
+				<a href="mailto:yongyetan1209@gmail.com" class="relative z-10" title="email">
+					<Gmail />
+				</a>
+				<div class="absolute top-1 -right-1 group-hover:block pt-6 pr-1 hidden">
+					<div class="bg-gray-200 px-2 py-0.5">
+						<p class="text-sm font-medium">yongyetan1209@gmail.com</p>
+					</div>
 				</div>
 			</div>
-		</div>
 
-        <div class="relative group">
-			<a href="https://github.com/algebra2boy" class="relative z-10" title="github"> 
-				<svg
-				xmlns="http://www.w3.org/2000/svg"
-				x="0px"
-				y="0px"
-				width="24"
-				height="24"
-				viewBox="0 0 100 100"
-			>
-				<circle cx="50" cy="50" r="30" fill="#4e6e91" /><path
-					fill="#6693c1"
-					d="M50,83c-18.196,0-33-14.804-33-33s14.804-33,33-33s33,14.804,33,33S68.196,83,50,83z M50,22 c-15.439,0-28,12.561-28,28s12.561,28,28,28s28-12.561,28-28S65.439,22,50,22z"
-				/><path
-					fill="#eeecd9"
-					d="M69.457,49.039c0-3.321-1.305-6.334-3.419-8.573c0.396-2.203,0.351-5.301-0.538-7.966 c-4.475,0-8.114,3.447-8.702,4H43.994c-0.589-0.552-4.019-4-8.494-4c-0.8,2.401-1.087,5.233-0.846,7.295 c-2.518,2.286-4.108,5.575-4.108,9.245c0,6.908,5.599,12.459,12.507,12.459H45.5c-2.003,0.917-3.635,2.756-4,5 c-2,0-4.864-0.182-6.181-2.158c-2.46-3.69-3.59-3.69-4.819-3.69c-1.23,0-1.33,1.23-0.1,2.46s1.23,1.23,2.46,3.69 C33.872,68.828,36.5,70.5,41.5,70.5v6.6c0,0,6.346,1.4,8.5,1.4s8.5-1.4,8.5-1.4l0-9.445c0-2.718-1.681-5.092-4-6.155h2.449 C63.858,61.5,69.457,55.947,69.457,49.039z"
-				/><path
-					fill="#1f212b"
-					d="M50,85c-19.299,0-35-15.701-35-35s15.701-35,35-35s35,15.701,35,35S69.299,85,50,85z M50,17 c-18.196,0-33,14.804-33,33s14.804,33,33,33s33-14.804,33-33S68.196,17,50,17z"
-				/><path
-					fill="#1f212b"
-					d="M50,79c-15.99,0-29-13.009-29-29s13.01-29,29-29s29,13.009,29,29c0,2.925-0.435,5.812-1.291,8.582 c-0.082,0.263-0.364,0.411-0.625,0.33c-0.264-0.082-0.412-0.361-0.33-0.625C77.581,55.612,78,52.825,78,50 c0-15.439-12.561-28-28-28S22,34.561,22,50s12.561,28,28,28c5.856,0,11.464-1.788,16.217-5.171c0.225-0.16,0.536-0.107,0.697,0.117 c0.16,0.225,0.107,0.537-0.117,0.697C61.873,77.147,56.065,79,50,79z"
-				/><path
-					fill="#1f212b"
-					d="M68.631,72.068c-0.14,0-0.279-0.059-0.378-0.173c-0.181-0.209-0.158-0.525,0.051-0.706 c0.739-0.638,1.452-1.324,2.122-2.037c0.188-0.202,0.505-0.21,0.706-0.023c0.201,0.189,0.212,0.505,0.023,0.707 c-0.693,0.739-1.433,1.449-2.197,2.11C68.863,72.028,68.747,72.068,68.631,72.068z"
-				/><path
-					fill="#1f212b"
-					d="M72.494,68.002c-0.107,0-0.216-0.035-0.308-0.105c-0.218-0.17-0.257-0.484-0.087-0.702 c1.649-2.118,2.982-4.452,3.963-6.938c0.101-0.258,0.392-0.382,0.648-0.282c0.257,0.102,0.383,0.392,0.281,0.648 c-1.015,2.575-2.396,4.993-4.104,7.186C72.79,67.936,72.643,68.002,72.494,68.002z"
-				/><path
-					fill="#1f212b"
-					d="M58.5,77.6c-0.276,0-0.5-0.224-0.5-0.5v-9.445c0-2.431-1.456-4.668-3.708-5.701 c-0.214-0.098-0.331-0.332-0.28-0.561C54.062,61.164,54.265,61,54.5,61h2.449c6.621,0,12.008-5.366,12.008-11.961 c0-3.064-1.166-5.987-3.282-8.229c-0.109-0.115-0.157-0.275-0.129-0.432c0.333-1.854,0.39-4.725-0.409-7.37 c-3.682,0.162-6.795,2.725-7.987,3.848C57.057,36.943,56.925,37,56.798,37H43.993c-0.127,0-0.249-0.048-0.342-0.135l-0.092-0.087 c-3.078-2.927-5.829-3.686-7.697-3.77c-0.662,2.162-0.941,4.762-0.712,6.729c0.019,0.16-0.042,0.319-0.161,0.428 c-2.506,2.275-3.943,5.51-3.943,8.875C31.046,55.635,36.433,61,43.053,61H45.5c0.235,0,0.438,0.164,0.488,0.394 c0.051,0.229-0.066,0.463-0.28,0.561c-1.974,0.904-3.397,2.676-3.715,4.625C41.954,66.822,41.745,67,41.5,67 c-2.333,0-5.191-0.271-6.598-2.38c-2.311-3.467-3.28-3.467-4.403-3.467c-0.14,0-0.385,0.023-0.448,0.178 c-0.086,0.206,0.031,0.756,0.702,1.428c1.287,1.287,1.311,1.335,2.554,3.82C34.409,68.785,37.319,70,41.5,70 c0.276,0,0.5,0.224,0.5,0.5v6.6c0,0.276-0.224,0.5-0.5,0.5S41,77.376,41,77.1v-6.105c-5.797-0.131-7.866-2.525-8.588-3.969 c-1.194-2.387-1.194-2.387-2.366-3.56c-0.869-0.869-1.213-1.81-0.919-2.518c0.209-0.505,0.709-0.795,1.372-0.795 c1.588,0,2.81,0.272,5.235,3.912c1.069,1.604,3.359,1.9,5.356,1.932c0.362-1.545,1.309-2.965,2.63-3.997h-0.668 c-7.172,0-13.007-5.813-13.007-12.959c0-3.562,1.485-6.988,4.084-9.442c-0.209-2.203,0.128-4.956,0.896-7.257 C35.094,32.138,35.285,32,35.5,32c2.036,0,5.192,0.696,8.692,4h12.409c1.427-1.311,4.847-4,8.898-4 c0.215,0,0.406,0.138,0.475,0.342c0.935,2.802,0.928,5.901,0.598,7.965c2.185,2.404,3.385,5.495,3.385,8.732 C69.957,56.186,64.122,62,56.949,62h-0.693C57.963,63.368,59,65.442,59,67.655V77.1C59,77.376,58.776,77.6,58.5,77.6z"
-				/><path
-					fill="#1f212b"
-					d="M34.238,45.97c-0.063,0-0.127-0.012-0.188-0.037c-0.256-0.104-0.379-0.396-0.274-0.651 c0.476-1.167,1.167-2.226,2.057-3.148c0.192-0.198,0.509-0.204,0.707-0.013c0.199,0.192,0.205,0.508,0.014,0.707 c-0.802,0.831-1.425,1.783-1.852,2.831C34.622,45.853,34.436,45.97,34.238,45.97z"
-				/><path
-					fill="#1f212b"
-					d="M33.66,51.021c-0.241,0-0.454-0.176-0.493-0.422c-0.08-0.51-0.121-1.034-0.121-1.558 c0-0.533,0.042-1.067,0.124-1.59c0.044-0.273,0.305-0.459,0.572-0.416c0.272,0.043,0.459,0.299,0.416,0.572 c-0.074,0.471-0.112,0.953-0.112,1.434c0,0.472,0.037,0.943,0.109,1.402c0.043,0.273-0.144,0.529-0.416,0.572 C33.713,51.019,33.687,51.021,33.66,51.021z"
-				/><path
-					fill="#1f212b"
-					d="M41.457,58.914c-0.028,0-0.057-0.002-0.085-0.007c-3.442-0.585-6.372-2.975-7.646-6.238 c-0.101-0.257,0.026-0.547,0.284-0.647c0.256-0.101,0.547,0.027,0.647,0.284c1.146,2.938,3.783,5.089,6.882,5.615 c0.271,0.046,0.455,0.305,0.409,0.577C41.907,58.741,41.696,58.914,41.457,58.914z"
-				/>
-			</svg>
-			</a>
-			<div class="absolute top-1 -right-1 group-hover:block pt-6 pr-1 hidden">
-				<div class="bg-gray-200 px-2 py-0.5">
-					<p class="text-sm font-medium">https://github.com/algebra2boy</p>
+			<div class="relative group">
+				<a href="https://github.com/algebra2boy" class="relative z-10" title="github">
+					<Github />
+				</a>
+				<div class="absolute top-1 -right-1 group-hover:block pt-6 pr-1 hidden">
+					<div class="bg-gray-200 px-2 py-0.5">
+						<p class="text-sm font-medium">https://github.com/algebra2boy</p>
+					</div>
 				</div>
 			</div>
-		</div>
 
-        <div class="relative group">
-			<a href="https://linkedin.com/in/yytan04" class="relative z-10" title="linkedin">
-				<svg
-				xmlns="http://www.w3.org/2000/svg"
-				x="0px"
-				y="0px"
-				width="24"
-				height="24"
-				viewBox="0 0 48 48"
-			>
-				<path
-					fill="#0288D1"
-					d="M42,37c0,2.762-2.238,5-5,5H11c-2.761,0-5-2.238-5-5V11c0-2.762,2.239-5,5-5h26c2.762,0,5,2.238,5,5V37z"
-				/><path
-					fill="#FFF"
-					d="M12 19H17V36H12zM14.485 17h-.028C12.965 17 12 15.888 12 14.499 12 13.08 12.995 12 14.514 12c1.521 0 2.458 1.08 2.486 2.499C17 15.887 16.035 17 14.485 17zM36 36h-5v-9.099c0-2.198-1.225-3.698-3.192-3.698-1.501 0-2.313 1.012-2.707 1.99C24.957 25.543 25 26.511 25 27v9h-5V19h5v2.616C25.721 20.5 26.85 19 29.738 19c3.578 0 6.261 2.25 6.261 7.274L36 36 36 36z"
-				/>
-			</svg>
-			</a>
-			<div class="absolute top-1 -right-1 group-hover:block pt-6 pr-1 hidden">
-				<div class="bg-gray-200 px-2 py-0.5">
-					<p class="text-sm font-medium">https://linkedin.com/in/yytan04</p>
+			<div class="relative group">
+				<a href="https://linkedin.com/in/yytan04" class="relative z-10" title="linkedin">
+					<LinkedIn />
+				</a>
+				<div class="absolute top-1 -right-1 group-hover:block pt-6 pr-1 hidden">
+					<div class="bg-gray-200 px-2 py-0.5">
+						<p class="text-sm font-medium">https://linkedin.com/in/yytan04</p>
+					</div>
 				</div>
 			</div>
 		</div>
