+++ src/routes/courses/+page.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+	import { Course } from '$lib/components/index.js';
+</script>
+
+<Course />
