+++ src/routes/courses/list/+page.svelte
