+++ src/routes/project/personal/+page.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+	import { Project } from '$lib/components/index.js';
+</script>
+
+<Project />
