+++ src/routes/project/personal/[name]/+page.server.ts
@@ -3,7 +3,7 @@ import { error } from '@sveltejs/kit';
 import type { PageServerLoad } from './$types';
 import type { PersonalProject } from '$lib/components/project/';
 
-import { personalProjects } from '$lib/components/project/personal/personalProjectData';
+import { personalProjects } from '$lib/components/project/personalProjectData';
 
 export const load: PageServerLoad = ({ params }) => {
 	const name = params.name;
